+++ src/constants.ts
@@ -11,18 +11,3 @@ export const productionEnvironmentName: string = 'Production';
 export const appSubpathSetting: string = 'appSubpath';
 export const apiSubpathSetting: string = 'apiSubpath';
 export const appArtifactSubpathSetting: string = 'appArtifactSubpath';
-
-// Doc for these parameter values: https://developer.github.com/v3/checks/runs/#parameters
-
-export enum Conclusion {
-    Success = 'success',
-    Failure = 'failure',
-    Skipped = 'skipped',
-    Cancelled = 'cancelled'
-}
-
-export enum Status {
-    Queued = 'queued',
-    InProgress = 'in_progress',
-    Completed = 'completed'
-}
