+++ src/gitHubTypings.ts
@@ -8,3 +8,18 @@ import { ActionsGetJobForWorkflowRunResponseData, OrgsListForAuthenticatedUserRe
 // we'll export the types until this is merged: https://github.com/octokit/types.ts/issues/120
 export type OrgForAuthenticatedUserData = OrgsListForAuthenticatedUserResponseData[0];
 export type ActionWorkflowStepData = ActionsGetJobForWorkflowRunResponseData['steps'][0];
+
+// Doc for these parameter values: https://developer.github.com/v3/checks/runs/#parameters
+
+export enum Conclusion {
+    Success = 'success',
+    Failure = 'failure',
+    Skipped = 'skipped',
+    Cancelled = 'cancelled'
+}
+
+export enum Status {
+    Queued = 'queued',
+    InProgress = 'in_progress',
+    Completed = 'completed'
+}
