+++ src/utils/gitHubUtils.ts
@@ -11,8 +11,8 @@ import { Response } from 'request';
 import * as git from 'simple-git/promise';
 import * as vscode from 'vscode';
 import { IAzureQuickPickItem } from 'vscode-azureextensionui';
-import { Conclusion, githubApiEndpoint, Status } from '../constants';
-import { OrgForAuthenticatedUserData } from '../gitHubTypings';
+import { githubApiEndpoint } from '../constants';
+import { Conclusion, OrgForAuthenticatedUserData, Status } from '../gitHubTypings';
 import { localize } from './localize';
 import { requestUtils } from './requestUtils';
 import { getSingleRootFsPath } from './workspaceUtils';
